Move DB subcommands to forest-tool
